% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitWD.R
\name{fitWD}
\alias{fitWD}
\title{Estimating parameters of the two-parameter Weibull distribution}
\usage{
fitWD(
  data,
  est.method,
  opt.method,
  starts,
  lower = NULL,
  upper = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{Numeric vector of observations.}

\item{est.method}{Character string specifying the estimation method.
Options include \code{"MLE"}, \code{"LSE"}, \code{"WLSE"}, and \code{"MPS"}.}

\item{opt.method}{Character string specifying the optimization method
used in \code{optim}, such as \code{"Nelder-Mead"}, \code{"BFGS"},
\code{"CG"}, \code{"L-BFGS-B"}, \code{"SANN"}, or \code{"Brent"}.}

\item{starts}{Numeric vector of initial values for the parameters}

\item{lower}{Numeric vector of lower bounds for parameters in constrained optimization.
Ignored if \code{NULL}.}

\item{upper}{Numeric vector of upper bounds for parameters in constrained optimization.}

\item{verbose}{Logical. If \code{TRUE}, prints optimization progress.}

\item{...}{Additional arguments passed to \code{optim}.}
}
\value{
A list containing:
\item{estimates}{Named numeric vector of estimated parameters \eqn{(a, b)}.}
\item{measures}{Numeric vector of model selection criteria (log-likelihood, AIC, BIC).}
\item{initials}{Initial values used in the optimization.}
\item{opt.fit}{Full output from \code{optim}.}
}
\description{
Estimates the parameters of the two-parameter Weibull distribution
using classical methods.
}
\details{
Classical estimations fo the parameters of the two-parameter Weibull distribution

The two-parameter Weibull Distribution has cumulative distribution
function (CDF) and probability density function (PDF):

\deqn{
F(x) = 1 - \exp(-a x^b),
}
\deqn{
f(x) = a b x^{b - 1} \exp(-a x^b ),
}

where \eqn{x > 0}, \eqn{a > 0} is the scale parameter and \eqn{b > 0} is the shape parameter.

The parameters are estimated using the following methods:

\itemize{
\item \strong{Maximum Likelihood Estimation (MLE):}
Maximizes the log-likelihood under the MWD model.

\item \strong{Least Squares Estimation (LSE):}
Minimizes squared differences between empirical and theoretical CDFs.
The empirical CDF uses Benard's approximation:
\eqn{F(x_{(i)}) = (i - 0.3)/(n + 0.4)}, for \eqn{i = 1, \dots, n}.

\item \strong{Weighted Least Squares Estimation (WLSE):}
A modification of LSE that assigns weights to the squared differences.
Uses weights
\eqn{w_i = \frac{(n+1)^2(n+2)}{i(n-i+1)}}, for \eqn{i = 1, \dots, n}.

\item \strong{Maximum Product of Spacings (MPS):}
Maximizes the product of spacings of the fitted CDF.
}
}
\examples{
# generate data from WD(a, b)
n <- 50
a <- 0.75; b <- 1.25; lambda <- 0 # reduces two-parameter Weibull distribution
set.seed(123)
X <- rMweibull(n, a, b, lambda)
init <- runif(2)

# fit model
fit.mle <- fitWD(data = X, est.method = "mle", opt.method = "L-BFGS-B", starts = init,
                 lower = c(1e-05,1e-05), upper = c(Inf,Inf), hessian = FALSE )
fit.mle$estimates

fit.lse <- fitWD(data = X, est.method = "lse", opt.method = "L-BFGS-B", starts = init,
                 lower = c(1e-05,1e-05), upper = c(Inf,Inf), hessian = FALSE )
fit.lse$estimates

fit.wlse <- fitWD(data = X, est.method = "wlse", opt.method = "L-BFGS-B", starts = init,
                  lower = c(1e-05,1e-05), upper = c(Inf,Inf), hessian = FALSE )
fit.wlse$estimates

fit.mps <- fitWD(data = X, est.method = "mps", opt.method = "L-BFGS-B", starts = init,
                 lower = c(1e-05,1e-05), upper = c(Inf,Inf), hessian = FALSE )
fit.mps$estimates

}
