% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitClayton.R
\name{mps_clayton}
\alias{mps_clayton}
\title{since the original obj function is maximized, it is multiplying with -1 for minimizing}
\usage{
mps_clayton(par, x, y, estimates)
}
\description{
since the original obj function is maximized, it is multiplying with -1 for minimizing
}
\keyword{internal}
