% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{Reliability_Clayton_MWD_MC}
\alias{Reliability_Clayton_MWD_MC}
\title{Monte Carlo Estimation of Stress–Strength Reliability (MWD + Clayton Copula)}
\usage{
Reliability_Clayton_MWD_MC(a1, b1, lambda1, a2, b2, lambda2, theta, N = 10000)
}
\arguments{
\item{a1, b1, lambda1}{Parameters of the strength variable \eqn{X},
with \eqn{a_1 > 0}, \eqn{b_1 \ge 0}, and \eqn{\lambda_1 \ge 0}.}

\item{a2, b2, lambda2}{Parameters of the stress variable \eqn{Y},
with \eqn{a_2 > 0}, \eqn{b_2 \ge 0}, and \eqn{\lambda_2 \ge 0}.}

\item{theta}{Clayton copula dependence parameter, \eqn{\theta > 0}.}

\item{N}{Integer. Number of Monte Carlo samples from \eqn{X \sim MWD(a_1,b_1,\lambda_1)}.}
}
\value{
A list containing:
\item{R}{Approximated reliability value based on \eqn{N} generated random samples.}
\item{R_MCsample}{Monte Carlo approximate of the reliability \eqn{R} based on \eqn{N} generated samples.}
\item{sample}{Generated random sample from \eqn{MWD(a_1, b_1, \lambda_1)} with \eqn{N} size.}
}
\description{
Computes an approximate value of the stress–strength reliability
\eqn{R = P(X > Y)} using Monte Carlo integration method under the modified Weibull
model with dependence induced by a Clayton copula.

It calculates an approximate value of \eqn{R} using the Monte Carlo integration method based
on a random generated MWD sample from \eqn{MWD(a_1, b_1, \lambda_1)}.
}
\details{
Monte Carlo Estimation of Stress–Strength Reliability under Clayton Copula

The approximate stress–strength reliability \eqn{R} is approximated via Monte
Carlo integration:

\deqn{
R \approx \frac{1}{N} \sum_{i=1}^{N}
T(x_i; \boldsymbol{\Omega_1}, \boldsymbol{\Omega_2}, \theta) = \tilde{R},
}

where \eqn{\boldsymbol{\Omega_1} = (a_1, b_1, \lambda_1)} and
\eqn{\boldsymbol{\Omega_2} = (a_2, b_2, \lambda_2)} denote the marginal parameter vectors.
The function

\deqn{
T(x; \boldsymbol{\Omega_1}, \boldsymbol{\Omega_2}, \theta)
= F_X(x)^{-(\theta + 1)}
\left( F_X(x)^{-\theta} + G_Y(x)^{-\theta} - 1 \right)^{-\left(\frac{1}{\theta}+1\right)},
}

where \eqn{x_1, \dots, x_N} is a random sample generated from the
Modified Weibull distribution \eqn{MWD(a_1, b_1, \lambda_1)}.

The accuracy of the approximation improves as sample size \eqn{N} increases.
In practice, values around \eqn{N = 10^5}  typically provide stable results.

Further details can be found in Kizilaslan (2026).
}
\examples{
# example code
a1 <- 0.75; b1 <- 1.5; lambda1 <- 0.6
a2 <- 1.2; b2 <- 0.5; lambda2 <- 0.9
theta <- 1 # 2, 3, 4, 5
R <- Reliability_Clayton_MWD(a1, b1, lambda1, a2, b2, lambda2, theta)
R$value
# approximated reliability R based on MC method
R_MC <- Reliability_Clayton_MWD_MC(a1, b1, lambda1, a2, b2, lambda2, theta, N = 50000)
R_MC$R

}
\references{
Kizilaslan, F. (2026).
\emph{Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins}.
\href{https://arxiv.org/abs/2604.12130}{arXiv:2604.12130}
}
