% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMweibull_Clayton.R
\name{rMweibull_Clayton}
\alias{rMweibull_Clayton}
\title{Random Generation for MWD Marginals via Clayton Copula}
\usage{
rMweibull_Clayton(n, a1, b1, lambda1, a2, b2, lambda2, theta)
}
\arguments{
\item{n}{Integer. Number of observations to be generated.}

\item{a1, b1, lambda1}{Parameters of the strength variable \eqn{X},
with \eqn{a_1 > 0}, \eqn{b_1 \ge 0}, and \eqn{\lambda_1 \ge 0}.}

\item{a2, b2, lambda2}{Parameters of the stress variable \eqn{Y},
with \eqn{a_2 > 0}, \eqn{b_2 \ge 0}, and \eqn{\lambda_2 \ge 0}.}

\item{theta}{Clayton copula dependence parameter, \eqn{\theta > 0}.}
}
\value{
A list containing:

A list containing:
\item{U}{Uniform samples used in the copula construction.}
\item{V}{Dependent uniform samples generated via the Clayton copula.}
\item{X}{Simulated observations from \eqn{X \sim \mathrm{MWD}(a_1, b_1, \lambda_1)} obtained by transforming \eqn{U}.}
\item{Y}{Simulated observations from \eqn{Y \sim \mathrm{MWD}(a_2, b_2, \lambda_2)} obtained by transforming \eqn{V}.}
}
\description{
Generates bivariate random samples from a dependent stress–strength model where
both marginals follow the Modified Weibull Distribution (MWD), and the dependence
structure between the variables is modeled using a Clayton copula.

Generates bivariate random samples from a dependent stress–strength model where
both marginals follow the Modified Weibull Distribution (MWD), and dependence
between variables is modeled using a Clayton copula.
}
\details{
Bivariate Random Data Generation under Clayton Copula with MWD Marginals

This function generates dependent uniform variables using the Clayton copula,
which are then transformed via inverse CDFs of the Modified Weibull marginals
to obtain \eqn{(X, Y)}.

Further details are provided in Kizilaslan (2026).
}
\examples{
set.seed(123)
n <- 50
a1 <- 0.75; b1 <- 1.5; lambda1 <- 0.6
a2 <- 1.2; b2 <- 0.5; lambda2 <- 0.9
theta <- 1 # 2, 3, 4, 5
# data generation
dat <- rMweibull_Clayton(n, a1, b1, lambda1, a2, b2, lambda2, theta)
str(dat)

}
\references{
Kizilaslan, F. (2026).
\emph{Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins}.
\href{https://arxiv.org/abs/2604.12130}{arXiv:2604.12130}
}
