% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDams.R
\docType{data}
\name{AllDams}
\alias{AllDams}
\title{All Istanbul Dams Data}
\format{
A data frame with 8520 rows and 13 variables.
}
\usage{
AllDams
}
\description{
Daily data for 10 dams in Istanbul, Türkiye. The dataset consists of daily
occupancy rates of Istanbul’s dams, retrieved in March 2026 from
\href{https://data.ibb.gov.tr/en/dataset/istanbul-barajlari-gunluk-doluluk-oranlari/resource/af0b3902-cfd9-4096-85f7-e2c3017e4f21}{https://data.ibb.gov.tr}.
The data span the period from late October 2000 to mid-February 2024.
}
\keyword{data}
