% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clayton.Copula.R
\name{Clayton_Copula}
\alias{Clayton_Copula}
\alias{Clayton_Copula_pdf}
\title{Two-dimensional Clayton Copula}
\usage{
Clayton_Copula(u, v, theta)

Clayton_Copula_pdf(u, v, theta)
}
\arguments{
\item{u}{is a vector of points in \eqn{[0,1]} representing the first coordinate of the copula.}

\item{v}{is a vector of points in \eqn{[0,1]} representing the first coordinate of the copula.}

\item{theta}{defines \eqn{\theta} dependency parameter between the pairs.}
}
\value{
\code{Clayton_Copula} gives the joint cumulative distribution function,
\code{Clayton_Copula_pdf} gives the joint probability density function values for the Clayton copula evaluated at points \eqn{(u,v)}
with given \eqn{\theta} parameter.
}
\description{
This is the joint cumulative distribution and probability density functions
for the two-dimensional Clayton copula.
}
\details{
The joint distribution function of the two-dimensional Clayton copula, along with its joint
probability density function, are given by
\deqn{C(u,v;\theta) = \left(u^{-\theta} + v^{-\theta} - 1\right)^{-1/\theta}, \; \theta \in (0, \infty).}
\deqn{c_{\theta}(u,v) = (\theta +1)  u^{-(\theta + 1)} v^{-(\theta + 1)} \left( u^{-\theta} + v^{-\theta}-1 \right)^{-\left (\frac{1}{\theta} + 2 \right)}}
where \eqn{\theta > 0}.
}
