% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.SSR.ClaytonMWD.R
\name{fit.SSR.ClaytonMWD}
\alias{fit.SSR.ClaytonMWD}
\title{Estimation of SSR Model Parameters with MWD Marginals via Clayton Copula}
\usage{
fit.SSR.ClaytonMWD(
  data,
  ACI = FALSE,
  bootstrap = FALSE,
  B = NULL,
  seed = NULL,
  one.step = TRUE,
  alpha = 0.05,
  verbose = FALSE
)
}
\arguments{
\item{data}{A list containing two numeric vectors:
\code{X} (strength) and \code{Y} (stress).}

\item{ACI}{Logical. If \code{TRUE}, asymptotic \eqn{95\%} confidence intervals
based on MLE are computed.}

\item{bootstrap}{Logical. If \code{TRUE}, parametric bootstrap confidence
intervals are computed.}

\item{B}{Integer. Number of bootstrap replications.}

\item{seed}{Integer. Random seed for reproducibility.}

\item{one.step}{Logical. If \code{TRUE},one-step LSE and WLSE estimators
are used for \eqn{\theta}.}

\item{alpha}{Numeric. Significance level for confidence intervals
(e.g., \code{0.05} for a \eqn{95\%} confidence interval).}

\item{verbose}{Logical; if \code{TRUE}, progress and intermediate
results from the optimization procedure are printed. Default is \code{FALSE}.}
}
\value{
A list containing:
\item{all.results}{Point estimates of all model parameters.}
\item{theta.Ktau}{Kendall's tau estimate corresponding to \eqn{\theta}.}
\item{seed}{Random seed used in the analysis.}
\item{data}{Input dataset used in the analysis.}
\item{ACI.parameters}{If \code{ACI = TRUE}, asymptotic \eqn{95\%} confidence intervals
for model parameters.}
\item{boot.mle}{If \code{bootstrap = TRUE}, bootstrap confidence intervals
based on MLE.}
\item{boot.lse}{If \code{bootstrap = TRUE}, bootstrap confidence intervals
based on LSE.}
\item{boot.wlse}{If \code{bootstrap = TRUE}, bootstrap confidence intervals
based on WLSE.}
\item{boot.mps}{If \code{bootstrap = TRUE}, bootstrap confidence intervals
based on MPS.}
\item{boot.samples}{A list containing bootstrap samples for all parameters
across all methods.}
}
\description{
Fits a dependent stress–strength reliability (SSR) model in which both
strength and stress follow the Modified Weibull Distribution (MWD), and
dependence is modeled using a Clayton copula.

The function estimates marginal parameters
\eqn{(a_1, b_1, \lambda_1)} for strength \eqn{X},
\eqn{(a_2, b_2, \lambda_2)} for stress \eqn{(Y},
and the copula dependence parameter \eqn{\theta}.

Estimation is performed using Maximum Likelihood Estimation (MLE),
Least Squares Estimation (LSE), Weighted Least Squares Estimation (WLSE),
and Maximum Product of Spacings (MPS).
}
\details{
Fit SSR Model with Modified Weibull Marginals via Clayton Copula

Returns point estimates and interval estimates of model parameters using
MLE, LSE, WLSE, and MPS methods.

Further theoretical details are available in Kizilaslan (2026).
}
\examples{
data <- list(X = TerkosDam, Y = OmerliDam)
fit.SSR <- fit.SSR.ClaytonMWD(data,  ACI = TRUE, bootstrap = TRUE, B = 10,
                             seed = 2026, one.step = TRUE, alpha = 0.05)
print(fit.SSR)

}
\references{
Kizilaslan, F. (2026).
\emph{Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins}.
\href{https://arxiv.org/abs/2604.12130}{arXiv:2604.12130}
}
