% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitClayton.R
\name{fitClayton}
\alias{fitClayton}
\title{Estimating parameter of the Clayton copula}
\usage{
fitClayton(
  x,
  y,
  est.method,
  opt.method,
  start,
  estimates,
  lower = NULL,
  upper = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{Vector of observations for the strength variable \eqn{X}.}

\item{y}{Vector of observations for the stress variable \eqn{Y}.}

\item{est.method}{Used method for estimating the parameters such as maximum likelihood estimate "MLE",
the least square estimation "LSE", the weighted least square estimation "WLSE", and
the maximum product of spacing estimates "MPS".}

\item{opt.method}{The optimization method for \code{optim} function such as "Nelder-Mead", "BFGS", "CG",
"L-BFGS-B", "SANN" and "Brent" to be used for estimating the parameters.}

\item{start}{Initial value for \eqn{\theta} parameter to be optimized over.}

\item{estimates}{A list containing estimates of all model paramters \eqn{a_1,b_1,\lambda_1, a_2, b_2, \lambda_2}.}

\item{lower}{Numeric vector specifying the lower bounds of the parameters
for bounded optimization methods (e.g., \code{"L-BFGS-B"}). Ignored if \code{NULL}.}

\item{upper}{Numeric vector specifying the upper bounds of the parameters
for bounded optimization methods. Ignored if \code{NULL}.}

\item{verbose}{Logical; if \code{TRUE}, progress and intermediate
results from the optimization procedure are printed. Default is \code{FALSE}.}

\item{...}{Further arguments to be passed to \code{optim} function such as lower and upper limits, hessian, etc.}
}
\value{
A list containing:
\item{estimate}{Estimate of the Clayton copula parameter, \eqn{\theta}.}
\item{opt.fit}{Full optimization result.}
}
\description{
Estimates the dependence parameter \eqn{\theta} of the Clayton copula.
}
\details{
The joint distribution function of the two-dimensional Clayton copula is
\deqn{
C(u,v;\theta) = \left(u^{-\theta} + v^{-\theta} - 1\right)^{-1/\theta},
}
where the dependence parameter satisfies \eqn{\theta > 0}.
It is estimated using several classical methods:

\itemize{
\item \strong{Maximum Likelihood Estimation (MLE):}
Obtains parameter estimates by maximizing the log-likelihood
function of the observed data under the assumed MWD distribution.

\item \strong{Least Squares Estimation (LSE):}
Estimates parameters by minimizing the sum of squared differences
between the empirical distribution function and theoretical distribution function
of MWD.
The Benard’s approximation is used for the empirical distribution function
at the ordered observations as \eqn{F(x_(i)) = (i-0.3)/(n+0.4), i=1,...,n.}.

\item \strong{Weighted Least Squares Estimation (WLSE):}
A modification of LSE that assigns weights to the squared differences.
The weights are taken as \eqn{w_i = ((n+1)^2(n+2)) / (i(n-i+1)), i=1,...,n. }

\item \strong{Maximum Product of Spacings (MPS):}
Estimates parameters by maximizing the product of spacings between
consecutive values of the fitted distribution function, providing
a robust alternative to MLE, particularly in small samples.
}
Further details can be found in Kızılaslan (2026).
}
\references{
Kızılaslan, F. (2026).
\emph{Reliability estimation in dependent stress--strength model with Clayton copula and modified Weibull margins}.
arXiv preprint. Available at
\href{https://arxiv.org/abs/????}{https://arxiv.org/abs/????}.
}
