% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMWD.R
\name{fitMWD}
\alias{fitMWD}
\title{Estimating parameters of the Modified Weibull Distribution (MWD)}
\usage{
fitMWD(
  data,
  est.method,
  opt.method,
  starts,
  lower = NULL,
  upper = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{Vector of observations.}

\item{est.method}{Used method for estimating the parameters such as maximum likelihood estimate "MLE",
the least square estimation "LSE", the weighted least square estimation "WLSE", and
the maximum product of spacing estimates "MPS".}

\item{opt.method}{The optimization method for \code{optim} function such as "Nelder-Mead", "BFGS", "CG",
"L-BFGS-B", "SANN" and "Brent" to be used for estimating the parameters.}

\item{starts}{Initial values for the parameters to be optimized over.}

\item{lower}{Numeric vector specifying the lower bounds of the parameters
for bounded optimization methods (e.g., \code{"L-BFGS-B"}). Ignored if \code{NULL}.}

\item{upper}{Numeric vector specifying the upper bounds of the parameters
for bounded optimization methods. Ignored if \code{NULL}.}

\item{verbose}{Logical; if \code{TRUE}, progress and intermediate
results from the optimization procedure are printed. Default is \code{FALSE}.}

\item{...}{Further arguments to be passed to \code{optim} function such as lower and upper limits, hessian, etc.}
}
\value{
A list containing:
\item{estimates}{Estimated values of the model parameters.}
\item{measures}{Model selection criteria, including the log-likelihood, AIC, and BIC,
evaluated at the estimated parameter values.}
\item{initials}{Initial values used in the optimization procedure.}
\item{opt.fit}{Full optimization output.}
}
\description{
Estimates the parameters of the Modified Weibull Distribution (MWD)
using classical methods.
}
\details{
Classical estimations for the parameters of the Modified Weibull Distribution (MWD)

The Modified Weibull Distribution (Lai et al., 2003) has cumulative distribution
function (CDF) and probability density function (PDF) given by

\deqn{
F(x) = 1 - \exp\left(-a x^b \exp(\lambda x)\right),
}
\deqn{
f(x) = a (b + \lambda x) x^{b - 1} \exp(\lambda x)
\exp\left(-a x^b \exp(\lambda x)\right),
}

where \eqn{x > 0}, \eqn{a > 0} is a scale parameter, \eqn{b \ge 0} is a shape parameter,
and \eqn{\lambda \ge 0} is a flexibility parameter controlling the growth rate of the hazard function.

The model parameters are estimated using several classical methods:

\itemize{
\item \strong{Maximum Likelihood Estimation (MLE):}
Obtains parameter estimates by maximizing the log-likelihood
function of the observed data under the assumed MWD distribution.

\item \strong{Least Squares Estimation (LSE):}
Estimates parameters by minimizing the sum of squared differences
between the empirical distribution function and theoretical distribution function
of MWD.
The Benard’s approximation is used for the empirical distribution function
at the ordered observations as \eqn{F(x_(i)) = (i-0.3)/(n+0.4), i=1,...,n.}.

\item \strong{Weighted Least Squares Estimation (WLSE):}
A modification of LSE that assigns weights to the squared differences.
The weights are taken as \eqn{w_i = ((n+1)^2(n+2)) / (i(n-i+1)), i=1,...,n. }

\item \strong{Maximum Product of Spacings (MPS):}
Estimates parameters by maximizing the product of spacings between
consecutive values of the fitted distribution function, providing
a robust alternative to MLE, particularly in small samples.
}
Further details can be found in Kızılaslan (2026).
}
\examples{
# generate data from MWD(a, b, lambda)
n <- 100
a <- 0.75; b <- 1.25; lambda <- 0.60
set.seed(123)
dat <- rMweibull(n, a, b, lambda)
init <- runif(3)

# Fit MWD to dat.
fit.mle <- fitMWD(data = dat, est.method = "mle", opt.method = "L-BFGS-B", starts = init,
                  lower = c(1e-05,1e-05,1e-05), upper = c(Inf,Inf,Inf), hessian = FALSE )
fit.mle$estimates

fit.lse <- fitMWD(data = dat, est.method = "lse", opt.method = "L-BFGS-B", starts = init,
                  lower = c(1e-05,1e-05,1e-05), upper = c(Inf,Inf,Inf), hessian = FALSE )
fit.lse$estimates

fit.wlse <- fitMWD(data = dat, est.method = "wlse", opt.method = "L-BFGS-B", starts = init,
                   lower = c(1e-05,1e-05,1e-05), upper = c(Inf,Inf,Inf), hessian = FALSE )
fit.wlse$estimates

fit.mps <- fitMWD(data = dat, est.method = "mps", opt.method = "L-BFGS-B", starts = init,
                  lower = c(1e-05,1e-05,1e-05), upper = c(Inf,Inf,Inf), hessian = FALSE )
fit.mps$estimates
  
}
\references{
Lai, C. D., Xie, M., and Murthy, D. N. P. (2003).
\href{A modified Weibull distribution.}{https://doi.org/10.1109/TR.2002.805788}
\emph{IEEE Transactions on Reliability}, \strong{52}(1), 33--37.

Kızılaslan, F. (2026).
\emph{Reliability estimation in dependent stress--strength model with Clayton copula and modified Weibull margins}.
arXiv preprint. Available at
\href{https://arxiv.org/abs/????}{https://arxiv.org/abs/????}.
}
