% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitWD.R
\name{fitWD}
\alias{fitWD}
\title{Estimating parameters of the two-parameter Weibull distribution}
\usage{
fitWD(
  data,
  est.method,
  opt.method,
  starts,
  lower = NULL,
  upper = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{Vector of observations.}

\item{est.method}{Used method for estimating the parameters such as maximum likelihood estimate "MLE",
the least square estimation "LSE", the weighted least square estimation "WLSE", and
the maximum product of spacing estimates "MPS".}

\item{opt.method}{The optimization method for \code{optim} function such as "Nelder-Mead", "BFGS", "CG",
"L-BFGS-B", "SANN" and "Brent" to be used for estimating the parameters.}

\item{starts}{Initial values for the parameters to be optimized over.}

\item{lower}{Numeric vector specifying the lower bounds of the parameters
for bounded optimization methods (e.g., \code{"L-BFGS-B"}). Ignored if \code{NULL}.}

\item{upper}{Numeric vector specifying the upper bounds of the parameters
for bounded optimization methods. Ignored if \code{NULL}.}

\item{verbose}{Logical; if \code{TRUE}, progress and intermediate
results from the optimization procedure are printed. Default is \code{FALSE}.}

\item{...}{Further arguments to be passed to \code{optim} function such as lower and upper limits, hessian, etc.}
}
\value{
A list containing:
\item{estimates}{Estimated values of the model parameters.}
\item{measures}{Model selection criteria, including the log-likelihood, AIC, and BIC,
evaluated at the estimated parameter values.}
\item{initials}{Initial values used in the optimization procedure.}
\item{opt.fit}{Full optimization output.}
}
\description{
Estimates the parameters of the two-parameter Weibull distribution
using classical methods.
}
\details{
Classical estimations fo the parameters of the two-parameter Weibull distribution

The two-parameter Weibull Distribution has cumulative distribution
function (CDF) and probability density function (PDF) given by

\deqn{F(x) = 1 - \exp(-a x^b),}
\deqn{f(x) = a b x^{b - 1} \exp(-a x^b ),}

where \eqn{x > 0}, \eqn{a > 0} is the scale parameter and \eqn{b > 0} is the shape parameter.

The model parameters are estimated using several classical methods:

\itemize{
\item \strong{Maximum Likelihood Estimation (MLE):}
Obtains parameter estimates by maximizing the log-likelihood
function of the observed data under the assumed two-parameter Weibull distribution.

\item \strong{Least Squares Estimation (LSE):}
Estimates parameters by minimizing the sum of squared differences
between the empirical distribution function and theoretical distribution function
of MWD.
The Benard’s approximation is used for the empirical distribution function
at the ordered observations as \eqn{F(x_(i)) = (i-0.3)/(n+0.4), i=1,...,n.}.

\item \strong{Weighted Least Squares Estimation (WLSE):}
A modification of LSE that assigns weights to the squared differences.
The weights are taken as \eqn{w_i = ((n+1)^2(n+2)) / (i(n-i+1)), i=1,...,n. }

\item \strong{Maximum Product of Spacings (MPS):}
Estimates parameters by maximizing the product of spacings between
consecutive values of the fitted distribution function, providing
a robust alternative to MLE, particularly in small samples.
}
}
\examples{
# generate data from WD(a, b)
n <- 50
a <- 0.75; b <- 1.25; lambda <- 0 # reduces two-parameter Weibull distribution
set.seed(123)
X <- rMweibull(n, a, b, lambda)
init <- runif(2)

# fit model
fit.mle <- fitWD(data = X, est.method = "mle", opt.method = "L-BFGS-B", starts = init,
                 lower = c(1e-05,1e-05), upper = c(Inf,Inf), hessian = FALSE )
fit.mle$estimates

fit.lse <- fitWD(data = X, est.method = "lse", opt.method = "L-BFGS-B", starts = init,
                 lower = c(1e-05,1e-05), upper = c(Inf,Inf), hessian = FALSE )
fit.lse$estimates

fit.wlse <- fitWD(data = X, est.method = "wlse", opt.method = "L-BFGS-B", starts = init,
                  lower = c(1e-05,1e-05), upper = c(Inf,Inf), hessian = FALSE )
fit.wlse$estimates

fit.mps <- fitWD(data = X, est.method = "mps", opt.method = "L-BFGS-B", starts = init,
                 lower = c(1e-05,1e-05), upper = c(Inf,Inf), hessian = FALSE )
fit.mps$estimates

}
