% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitClayton.R
\name{LSE_clayton_onestep}
\alias{LSE_clayton_onestep}
\title{One-step LSE estimate of the Clayton copula parameter}
\usage{
LSE_clayton_onestep(par, x, y, estimates)
}
\arguments{
\item{par}{Numeric value of the dependence parameter \eqn{\theta} of the Clayton copula.}

\item{x}{Vector of observations for the strength variable \eqn{X}.}

\item{y}{Vector of observations for the stress variable \eqn{Y}.}

\item{estimates}{A list containing estimates of the model parameters
\eqn{a_1, b_1, \lambda_1, a_2, b_2, \lambda_2}.}
}
\value{
The one-step LSE of \eqn{\theta} parameter.
}
\description{
One-step LSE estimate of the Clayton copula parameter
}
\details{
Further details are provided in Kızılaslan (2026).
}
\references{
Kızılaslan, F. (2026).
\emph{Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins}.
\href{https://arxiv.org/abs/????}{arXiv:????}
}
