% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MWD.R
\name{ModifiedWeibull}
\alias{ModifiedWeibull}
\alias{dMweibull}
\alias{pMweibull}
\alias{qMweibull}
\alias{rMweibull}
\alias{hMweibull}
\title{The Modified Weibull Distribution (MWD)}
\usage{
dMweibull(x, a, b, lambda, log = FALSE)

pMweibull(q, a, b, lambda, lower.tail = TRUE, log = FALSE)

qMweibull(p, a, b, lambda, lower.tail = TRUE)

rMweibull(n, a, b, lambda)

hMweibull(x, a, b, lambda, log = FALSE)
}
\arguments{
\item{x}{Numeric vector of observations.}

\item{a}{Positive scale parameter of the distribution (\eqn{a > 0}).}

\item{b}{Non-negative shape parameter of the distribution (\eqn{b \ge 0}).}

\item{lambda}{Non-negative acceleration or flexibility parameter (\eqn{\lambda \ge 0})
that governs how quickly the risk of failure (hazard) grows over time.
Larger values correspond to faster failure accumulation and higher fragility,
and restricting \eqn{\lambda \ge 0} ensures a non-decreasing hazard function.}

\item{log}{Logical; if \code{TRUE}, then \eqn{log(f(t))} or \eqn{log(F(t))} or
\eqn{log(h(t))} are returned.}

\item{q}{Numeric vector of quantiles.}

\item{lower.tail}{Logical; if \code{FALSE}, then \eqn{1-F(t)} are returned and quantiles
are computed for \eqn{1-p}.}

\item{p}{Numeric vector of probabilities in \eqn{[0,1]}.}

\item{n}{Integer; number of observations to be generated.}
}
\value{
\code{dMweibull} gives the probability density function, \code{pMweibull} gives the cumulative distribution function,
\code{qMweibull} gives the quantile function, \code{rMweibull} generates random deviates, and
\code{hMweibull} gives the hazard function.
}
\description{
This is the probability density function, cumulative distribution function, quantile function,
random generation and hazard function for the Modified Weibull distribution (MWD) by Lai et al. (2003).
}
\details{
The Modified Weibull distribution with parameters \eqn{a},
\eqn{b} and \eqn{\lambda} has the cumulative distribution
density and hazard functions given by
\deqn{F(x) = 1 - \exp(-a x^b \exp(\lambda x)),}
\deqn{f(x) = a (b+ \lambda x) x^{b - 1} \exp(\lambda x) \exp(-a x^b \exp(\lambda x)),}
\deqn{h(x) = a (b+ \lambda x) x^{b - 1} \exp(\lambda x), }
where \eqn{x > 0}, \eqn{a > 0} is the scale parameter, \eqn{b \ge 0} is a shape parameter,
and \eqn{\lambda \ge 0} is an acceleration or flexibility parameter that controls
how quickly the hazard grows over time.
When \eqn{\lambda = 0}, the MWD reduces to the two-parameter Weibull distribution
\eqn{F(x) = 1 - \exp(-a x^b)} with the scale parameter \eqn{a} and the shape parameter \eqn{b}.
When \eqn{b = 0}, the MWD reduces to a type I extreme-value (or log-gamma) distribution with
\eqn{F(x) = 1 - \exp(-a \exp(\lambda x)) }.
}
\examples{
n <- 25
a <- 0.75; b <- 1.25; lambda <- 0.60
set.seed(123)
x <- rlnorm(n)
ff <- dMweibull(x, a, b, lambda)
FF <- pMweibull(x, a, b, lambda)
qq <- qMweibull(runif(n), a, b, lambda)
dat <- rMweibull(n, a, b, lambda)
hf <- hMweibull(x, a, b, lambda )

}
\references{
Lai, C. D., Xie, M., and Murthy, D. N. P. (2003).
\href{A modified Weibull distribution.}{https://doi.org/10.1109/TR.2002.805788}
\emph{IEEE Transactions on Reliability}, \strong{52}(1), 33--37.
}
