% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{parametric_bootstrap}
\alias{parametric_bootstrap}
\title{Parametric Bootstrap Confidence Intervals}
\usage{
parametric_bootstrap(
  est.method,
  opt.method,
  boot.estimates,
  n,
  B = 1000,
  seed = NULL,
  one.step = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{est.method}{Used method for estimating the parameters such as maximum likelihood estimate "MLE",
the least square estimation "LSE", the weighted least square estimation "WLSE", and
the maximum product of spacing estimates "MPS".}

\item{opt.method}{The optimization method for \code{optim} function such as "Nelder-Mead", "BFGS", "CG",
"L-BFGS-B", "SANN" and "Brent" to be used for estimating the parameters.}

\item{boot.estimates}{A list containing the estimates of the unknown parameters.
The elements \eqn{(a_1, b_1, \lambda_1)} correspond to the strength variable,
\eqn{(a_2, b_2, \lambda_2)} correspond to the stress variable, and \eqn{\theta} corresponds to Clayton copula parameter.}

\item{n}{sample size}

\item{B}{Number of bootstrap samples.}

\item{seed}{Integer seed for reproducibility.}

\item{one.step}{Logical; if TRUE, one-step LSE and WLSE methods are used for estimating \eqn{\theta}.}

\item{alpha}{Numeric; significance level for confidence intervals (e.g., \eqn{0.05} for \eqn{95\%} CI).}
}
\value{
A list containing:
\item{parameters.quantiles}{It gives the \eqn{100(1-\alpha)\%} lower and upper bounds.}
\item{boot.results}{It represents all the estimates values of the parameters through B bootstrap samples.}
}
\description{
It gives the parametric bootstrap confidence interval estimation for all the unknown parameters
along with the reliability \eqn{R} based on MLE, LSE, WLSE and MPS methods.
}
\details{
The parametric bootstrap percentile method is employed to be able to construct bootstrap confidence intervals
for unknown parameters and \eqn{R} as well based on MLE, LSE, WLSE and MPSE methods.
Further details can be found in Kızılaslan (2026).
}
\references{
Kızılaslan, F. (2026).
\emph{Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins}.
\href{https://arxiv.org/abs/????}{arXiv:????}
}
