% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SSRfit.R
\name{print.SSRfit}
\alias{print.SSRfit}
\title{Print SSR.ClaytonMWD Fit Results}
\usage{
\method{print}{SSRfit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{SSRfit} obtained from
\code{fit.SSR.ClaytonMWD}.}

\item{...}{Additional arguments passed to the print method.
For example, \code{digits} controls the number of decimal places
used when printing numerical results.}
}
\description{
Prints the results of an object of class \code{SSRfit} created by
\code{fit.SSR.ClaytonMWD}, including separate tables of parameter estimates
and interval estimates (MLE, LSE, WLSE, and MPS) for the model parameters,
the Clayton copula parameter \eqn{\theta}, and the reliability \eqn{R}.
}
\examples{
data = list(X = TerkosDam, Y = OmerliDam)
fit.SSR = fit.SSR.ClaytonMWD(data, ACI = TRUE, bootstrap = TRUE, B = 10,
                             seed = 2026, one.step = TRUE, alpha = 0.05)
print(fit.SSR)
print(fit.SSR, 3)
}
