% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{Reliability_Clayton_MWD_MC}
\alias{Reliability_Clayton_MWD_MC}
\title{Monte Carlo Estimation of Stress-Strength Reliability for MWD under Clayton Copula}
\usage{
Reliability_Clayton_MWD_MC(a1, b1, lambda1, a2, b2, lambda2, theta, N = 10000)
}
\arguments{
\item{a1, b1, lambda1}{MWD parameters for the strength variable \eqn{X},
with \eqn{a_1 > 0}, \eqn{b_1 \ge 0}, and \eqn{\lambda_1 \ge 0}.}

\item{a2, b2, lambda2}{MWD parameters for the stress variable \eqn{Y},
with \eqn{a_2 > 0}, \eqn{b_2 \ge 0}, and \eqn{\lambda_2 \ge 0}.}

\item{theta}{Clayton copula parameter with \eqn{\theta >0}.}

\item{N}{Generated sample size from \eqn{X \sim MWD(a_1,b_1,\lambda_1)} for Monte Carlo Integration method.}
}
\value{
A list containing:
\item{R}{Approximated reliability value based on \eqn{N} generated random samples.}
\item{R_MCsample}{Monte Carlo approximate of the reliability \eqn{R} based on \eqn{N} generated samples.}
\item{sample}{Generated random sample of size \eqn{N} from \eqn{MWD(a_1, b_1, \lambda_1)}.}
}
\description{
It calculates an approximate value of \eqn{R} using the Monte Carlo integration method based
on a random generated MWD sample from \eqn{MWD(a_1, b_1, \lambda_1)}.
}
\details{
The approximate SSR \eqn{R} is computed using Monte Carlo integration:

\deqn{
R \approx \frac{1}{N} \sum_{i=1}^{N}
T(x_i; \bm{\Omega_1}, \bm{\Omega_2}, \theta) = \tilde{R},
}

where \eqn{\bm{\Omega_1} = (a_1, b_1, \lambda_1)} and
\eqn{\bm{\Omega_2} = (a_2, b_2, \lambda_2)} denote the marginal parameter vectors.
The function

\deqn{
T(x; \bm{\Omega_1}, \bm{\Omega_2}, \theta)
= F_X(x)^{-(\theta + 1)}
\left( F_X(x)^{-\theta} + G_Y(x)^{-\theta} - 1 \right)^{-\left(\frac{1}{\theta}+1\right)},
}

where \eqn{x_1, \dots, x_N} is a random sample generated from the
Modified Weibull distribution \eqn{MWD(a_1, b_1, \lambda_1)}.

The accuracy of this approximation improves with increasing sample size \eqn{N}.
In practice, values around \eqn{N = 10^5} provide accurate numerical results.
Further details can be found in Kızılaslan (2026).
}
\examples{
# example code
a1 <- 0.75; b1 <- 1.5; lambda1 <- 0.6
a2 <- 1.2; b2 <- 0.5; lambda2 <- 0.9
theta <- 1 # 2, 3, 4, 5

R = Reliability_Clayton_MWD(a1, b1, lambda1, a2, b2, lambda2, theta)
R$value

R_MC = Reliability_Clayton_MWD_MC(a1, b1, lambda1, a2, b2, lambda2, theta, N = 100000)
R_MC$R

}
\references{
Kizilaslan, F. (2026).
\emph{Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins}.
\href{https://arxiv.org/abs/2604.12130}{arXiv:2604.12130}
}
