% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rMweibull_Clayton.R
\name{rMweibull_Clayton}
\alias{rMweibull_Clayton}
\title{Bivariate Random Data Generation for MWD Marginals via Clayton Copula}
\usage{
rMweibull_Clayton(n, a1, b1, lambda1, a2, b2, lambda2, theta)
}
\arguments{
\item{n}{Integer; number of observations to be generated.}

\item{a1, b1, lambda1}{MWD parameters for the strength variable \eqn{X},
with \eqn{a_1 > 0}, \eqn{b_1 \ge 0}, and \eqn{\lambda_1 \ge 0}.}

\item{a2, b2, lambda2}{MWD parameters for the stress variable \eqn{Y},
with \eqn{a_2 > 0}, \eqn{b_2 \ge 0}, and \eqn{\lambda_2 \ge 0}.}

\item{theta}{Clayton copula parameter with \eqn{\theta >0}.}
}
\value{
A list containing:
\item{(U, V)}{Independent generated uniform numbers for \eqn{(X,Y)}.}
\item{(X, Y)}{Dependent generated \eqn{n} pairs \eqn{(X,Y)} observations.}
}
\description{
Generates bivariate random samples from a dependent stress–strength model where
both marginals follow the Modified Weibull Distribution (MWD), and the dependence
structure between the variables is modeled using a Clayton copula.
}
\details{
Further details are provided in Kızılaslan (2026).
}
\examples{
set.seed(123)
n <- 50
a1 <- 0.75; b1 <- 1.5; lambda1 <- 0.6
a2 <- 1.2; b2 <- 0.5; lambda2 <- 0.9
theta <- 1 # 2, 3, 4, 5
# data generation
dat <- SSReliabilityClaytonMWD::rMweibull_Clayton(n, a1, b1, lambda1, a2, b2, lambda2, theta)
str(dat)

}
\references{
Kizilaslan, F. (2026).
\emph{Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins}.
\href{https://arxiv.org/abs/2604.12130}{arXiv:2604.12130}
}
