% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitClayton.R
\name{theta_Ktau_estimate}
\alias{theta_Ktau_estimate}
\title{Kendall's Tau Estimate of Clayton copula parameter}
\usage{
theta_Ktau_estimate(data)
}
\arguments{
\item{data}{A list with two numeric vectors: \eqn{X} (strength) and \eqn{Y} (stress).}
}
\value{
Moment estimate of \eqn{\theta} parameter based on the Kendall's \eqn{\tau}.
}
\description{
Kendall's Tau Estimate of Clayton copula parameter theta
}
\examples{
set.seed(123)
n <- 50
a1 <- 0.75; b1 <- 1.5; lambda1 <- 0.6
a2 <- 1.2; b2 <- 0.5; lambda2 <- 0.9
theta <- 5 # 1, 2, 3, 4, 5
# data generation
dat <- SSReliabilityClaytonMWD::rMweibull_Clayton(n, a1, b1, lambda1, a2, b2, lambda2, theta)
theta_Ktau_estimate(dat)

}
